+++ gwcelery/tasks/raven.py
@@ -5,15 +5,18 @@ from ligo.raven import gracedb_events
 
 from ..import app
 from . import gracedb
+from . import ligo_fermi_skymaps
 
 
 tl_cbc, th_cbc = -5, 1
 tl_burst, th_burst = -600, 60
 
 
-def calculate_coincidence_far(gracedb_id, group):
-    """Compute FAR for external trigger and superevent coincidence by calling
-    ligo.raven.search.calc_signif_gracedb.
+def calculate_spacetime_coincidence_far(gracedb_id, group):
+    """Compute spatio-temporal coincidence FAR for GRB external trigger and
+    superevent coincidence by calling ligo.raven.search.calc_signif_gracedb.
+    Note: this will only run if skymaps from both triggers are available to
+    download.
 
     Parameters
     ----------
@@ -23,7 +26,39 @@ def calculate_coincidence_far(gracedb_id, group):
         CBC or Burst; group of the preferred_event associated with the
         gracedb_id superevent
     """
+    preferred_skymap = ligo_fermi_skymaps.get_preferred_skymap(gracedb_id)
+    se = gracedb_events.SE(gracedb_id, fitsfile=preferred_skymap,
+                           gracedb=gracedb.client)
+    em_events = gracedb.get_superevent(gracedb_id)['em_events']
 
+    if group == 'CBC':
+        tl, th = tl_cbc, th_cbc
+    elif group == 'Burst':
+        tl, th = tl_burst, th_burst
+
+    canvas = chain()
+    for exttrig_id in em_events:
+        if gracedb.download('glg_healpix_all_bn_v00.fit', exttrig_id):
+            exttrig = gracedb_events.ExtTrig(exttrig_id,
+                                             gracedb=gracedb.client)
+            canvas |= (
+                calc_signif.si(se, exttrig, tl, th, incl_sky=True))
+
+    return canvas
+
+
+def calculate_coincidence_far(gracedb_id, group):
+    """Compute temporal coincidence FAR for external trigger and superevent
+    coincidence by calling ligo.raven.search.calc_signif_gracedb.
+
+    Parameters
+    ----------
+    gracedb_id: str
+        ID of the superevent trigger used by GraceDb
+    group: str
+        CBC or Burst; group of the preferred_event associated with the
+        gracedb_id superevent
+    """
     se = gracedb_events.SE(gracedb_id, gracedb=gracedb.client)
     em_events = gracedb.get_superevent(gracedb_id)['em_events']
 
@@ -38,16 +73,16 @@ def calculate_coincidence_far(gracedb_id, group):
             exttrig = gracedb_events.ExtTrig(exttrig_id,
                                              gracedb=gracedb.client)
             canvas |= (
-                calc_signif.si(se, exttrig, tl, th))
+                calc_signif.si(se, exttrig, tl, th, incl_sky=False))
 
     return canvas
 
 
 @app.task(shared=False)
-def calc_signif(se, exttrig, tl, th):
+def calc_signif(se, exttrig, tl, th, incl_sky):
     """Calculate FAR of GRB exttrig-GW coincidence"""
     return ligo.raven.search.calc_signif_gracedb(se, exttrig, tl, th,
-                                                 incl_sky=False)
+                                                 incl_sky=incl_sky)
 
 
 def coincidence_search(gracedb_id, alert_object, group=None, pipelines=[]):